+++ lib/ahl_scraper/resources/game.rb
@@ -344,6 +344,7 @@ module AhlScraper
       irregular_game_status = IRREGULAR_GAMES.dig(game_id.to_s, :status)
       return irregular_game_status if irregular_game_status
 
+      return "postponed" if @raw_data[:details][:status] == "Postponed"
       # return "forfeited" if game is a forfeit, need to figure this out if it happens
 
       return "finished" if @raw_data[:details][:final] == "1"