+++ spec/ahl_scraper/fetchers/scoreboards/data_fetcher_spec.rb
@@ -12,8 +12,8 @@ RSpec.describe AhlScraper::Scoreboards::DataFetcher do
     it "fetches scoreboard data for date range", :vcr do
       games = described_class.new(start_date: start_date, end_date: end_date).call
 
-      games.each do |_game|
-        expect(Date.parse(card[:Date])).to be_between(start_date, end_date)
+      games.each do |game|
+        expect(Date.parse(game[:Date])).to be_between(start_date, end_date)
       end
     end
   end